Add support for updating existing courses in EloquentCourseRepository
By trying to save an existing course with the previous version of `save`, an integrity exception was thrown. This could have been solved by telling Eloquent that the record already exists with `$model->exists = true;` but by doing so, we would break inserts.

To fix the issue without having to write separated methods for **insert**s and **update**s, we used Eloquent's `updateOrCreate`.